+++ CHANGELOG.md
@@ -15,6 +15,9 @@ For updates follow [@implydata](https://twitter.com/implydata) on Twitter.
 - Changed `Dataset#toJS` to return an object rather than just the `data` array
 - Datasets now actively maintain their `keys` array
 - `Dataset#flatten` now returns a new (flat) `Dataset` and thus the `parentName` option is no longer supported
+- In `Dataset#flatten`, `Dataset#ToCSV`, e.t.c `orderedColumns` option is no longer supported use `.select()` instead
+- `Dataset#getColumns` is now just `return this.flatten(options).attributes;`
+- `Dataset#getNestedColumns` was removed
 - Added `ThenExpression`